# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API document \u003cstyle\u003e.models {display: none !important}\u003c/style\u003e",
        "title": "COJ Specialist API",
        "termsOfService": "COJ",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ข้อมูลตนเอง",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "ข้อมูลตนเอง",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "สมัครเข้าใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "สมัครเข้าใช้งาน",
                "parameters": [
                    {
                        "description": "สมัครเข้าใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagCreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RegisterUser": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.SwagCreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Created Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.SwagError400": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Bad Request"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError404": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Not Found"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError500": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.SwagGetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.SwagUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "รูป Avatar",
                    "type": "string",
                    "example": "user01.png"
                },
                "email": {
                    "description": "อีเมล์",
                    "type": "string",
                    "example": "user01@email.com"
                },
                "firstName": {
                    "description": "ชื่อ",
                    "type": "string",
                    "example": "john"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "lastName": {
                    "description": "นามสกุล",
                    "type": "string",
                    "example": "doe"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string",
                    "example": "user01"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}